+++ src/components/Footer.js
@@ -1,0 +1,12 @@
+import './Fotter.css';
+
+function Footer() {
+    return(
+        <div className='main'>
+            <p>Made with 🖤 Tadiwanshe Shangwa 2022 ©</p>
+            <p className='versionDetailts'>Last updated 16 August 2022</p>
+        </div>
+    )
+}
+
+export default Footer
