+++ src/components/Fotter.css
@@ -1,0 +1,19 @@
+:root {
+    --primary: #FF5B15;
+    --secondary: #fff;
+    --background: rgb(19, 19, 19);
+  } 
+
+  /* Main container for footer */
+.main {
+    height: 10%;
+    border-radius: none;
+    width: 100%;
+    background-color: var(--primary);
+    margin-top: 30px;
+    /* font-size: 10px; */
+}
+
+.versionDetailts {
+    font-size: 10px;
+}
