+++ src/components/Nav.js
@@ -1,0 +1,9 @@
+const heading = `<Tadiwa Dev!/>`
+
+export default function NavBar() {
+    return(
+        <div className='nav'>
+          <h1 className='nav-logo'>{heading}</h1>
+        </div>
+    )
+}
