+++ src/pages/App.css
@@ -1,19 +1,23 @@
 :root {
-  --orange: #FF5B15;
+  --primary: #FF5B15;
+  --secondary: #fff;
+  --background: rgb(19, 19, 19);
 } 
 
 .App {
+  background-color: var(--background);
   text-align: center;
   display: flex;
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  color: #fff;
 }
 
 /* Navbar */
 .nav {
   width: 100%;
-  background-color: var(--orange);
+  background-color: var(--primary);
   height: 60px;
   display: flex;
   flex-direction: row;
@@ -23,7 +27,6 @@
 }
 
 .nav-logo {
-  color: rgb(23, 23, 23);
   font-size: 24px;
   font-weight: 300;
 }
@@ -52,6 +55,7 @@
 }
 
 .socials {
+  border-radius: 5px;
   width: 100%;
   height: 50px;
   display: flex;
@@ -60,7 +64,7 @@
   align-items: center;
   text-decoration: none;
   background-color: #FF5B15;
-  margin: 2px;
+  margin: 5px;
   color: white;
   transition: all, 0.4s;
 }
@@ -71,4 +75,15 @@
 
 .socials:focus {
   background-color: black;
+}
+
+/* For big screens */
+@media only screen and (min-width: 600px) {
+  .socials {
+    width: 50%;
+  }
+
+  .aboutTextWrap {
+    width: 50%;
+  }
 }
